Fixing aggregate null error and renaming variables

+++ src/main/java/io/github/notoriousnapper/pqservice/model/Aggregate.java
@@ -12,7 +12,7 @@ public class Aggregate {
     Move move;
 
     @JsonProperty("aggregateValue")
-    Integer aggregateValue;
+    Double aggregateValue;
 
     @JsonProperty("aggregateType")
     String aggregateType;
